lo_mas_top_website/requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.3
billiard==3.3.0.23
celery==3.1.18
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.2
django-background-tasks==1.2.5
django-compat==1.0.15
django-heroku==0.3.1
get==2019.4.13
gunicorn==20.0.4
idna==2.9
kombu==3.0.37
Pillow==7.0.0
post==2019.4.13
psycopg2==2.8.4
public==2019.4.13
pytz==2019.3
query-string==2019.4.13
redis==2.10.3
request==2019.4.13
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
